Move biweight_* functions to separate module
